+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:givelivly_beta/Screens/HomeScreen.dart';
-import 'package:givelivly_beta/Screens/LoginScreen.dart';
-import 'package:givelivly_beta/Routes/Routes.dart';
-import 'package:givelivly_beta/Screens/WelcomeScreen.dart';
+import 'package:givelivly_beta/Screens/home_screen.dart';
+import 'package:givelivly_beta/Screens/login_screen.dart';
+import 'package:givelivly_beta/Routes/routes.dart';
+import 'package:givelivly_beta/Screens/welcome_screen.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 void main() => runApp(const GiveLivly());
@@ -22,8 +22,8 @@ class GiveLivly extends StatelessWidget {
       )),
       routes: {
         Routes.welcomeRoute: (context) => const WelcomeScreen(),
-        Routes.homeRoute: (context) => const Home_screen(),
-        Routes.loginRoute: (context) => const Login_page(),
+        Routes.homeRoute: (context) => const HomeScreen(),
+        Routes.loginRoute: (context) => const LoginScreen(),
       },
     );
   }
